requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
faiss-cpu==1.13.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
narwhals==2.11.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
six==1.17.0
sympy==1.14.0
ta==0.11.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0

# Google Sheets integration
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
